% Auto-generated: do not edit by hand
\name{dashPaneSplit}

\alias{dashPaneSplit}

\title{DashPaneSplit component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dashPaneSplit(id=NULL, containerStyle=NULL, mainChildren=NULL,
mainStyle=NULL, panelOrder=NULL, resizerSize=NULL,
sidebarChildren=NULL, sidebarSize=NULL, sidebarStyle=NULL,
splitMode=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{containerStyle}{}

\item{mainChildren}{A list of or a singular dash component, string or number. }

\item{mainStyle}{}

\item{panelOrder}{A value equal to: 'mainfirst', 'sidebarfirst'. }

\item{resizerSize}{Numeric. }

\item{sidebarChildren}{A list of or a singular dash component, string or number. }

\item{sidebarSize}{Character | numeric. }

\item{sidebarStyle}{}

\item{splitMode}{A value equal to: 'vertical', 'horizontal'. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

